Moved interfaces to contracts dir
