+++ services/assistant/app/contracts/clients/booking_client.py
