+++ services/assistant/app/contracts/repositories/__init__.py
@@ -1,0 +1,5 @@
+from app.contracts.repositories.booking_repository import BookingRepository
+from app.contracts.repositories.xray_repository import XrayRepository
+from app.contracts.repositories.user_repository import UserRepository
+
+__all__ = ["BookingRepository", "XrayRepository", "UserRepository"]
