+++ services/assistant/app/contracts/repositories/booking_repository.py
