+++ services/assistant/app/contracts/repositories/user_repository.py
