+++ services/assistant/app/contracts/repositories/xray_repository.py
