--- services/assistant/app/repositories/__init__.py
@@ -1,5 +1,0 @@
-from app.repositories.booking_repository import BookingRepository
-from app.repositories.xray_repository import XrayRepository
-from app.repositories.user_repository import UserRepository
-
-__all__ = ["BookingRepository", "XrayRepository", "UserRepository"]
